+++ apps/create-react-app/src/App.tsx
@@ -11,6 +11,22 @@ function App() {
       <button onClick={() => stop()}>Stop</button>
       <button onClick={pause}>Pause</button>
       <button onClick={resume}>Resume</button>
+      <button
+        onClick={() => {
+          start(0, 0, true);
+          window.history.pushState(null, '', '/?test=1');
+        }}
+      >
+        push url in param (disable auto stop)
+      </button>
+      <button
+        onClick={() => {
+          start(0, 0, false);
+          window.history.pushState(null, '', '/?test=1');
+        }}
+      >
+        push url in param (enable auto stop)
+      </button>
     </div>
   );
 }