+++ apps/create-react-app/src/ProgressRouter.tsx
@@ -1,0 +1,12 @@
+import { useAnchorProgress } from '@bprogress/react';
+
+export const ProgressRouter = () => {
+  useAnchorProgress(
+    {
+      forcedStopDelay: 500,
+    },
+    [],
+  );
+
+  return null;
+};