+++ apps/create-react-app/src/index.tsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import { ProgressProvider } from '@bprogress/react';
+import { ProgressRouter } from './ProgressRouter';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
@@ -10,6 +11,7 @@ const root = ReactDOM.createRoot(
 root.render(
   <React.StrictMode>
     <ProgressProvider>
+      <ProgressRouter />
       <App />
     </ProgressProvider>
   </React.StrictMode>,