+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bprogress/react
 
+## 1.2.5
+
+### Patch Changes
+
+- feat: add auto stop control
+
 ## 1.2.4
 
 ### Patch Changes